EEAIAdmin/requirements_v2.txt
# Original dependencies (preserved for compatibility)
oracledb~=2.5.1
sqlalchemy~=2.0.37
flask~=3.1.0
python-dotenv~=1.0.1
PyMuPDF==1.24.0
numpy~=1.25.2
faiss-cpu~=1.10.0
requests~=2.32.3
pandas~=2.2.3
pdfplumber~=0.11.5
pypdf~=5.3.0
PyPDF2~=3.0.1
azure-core~=1.32.0
beautifulsoup4~=4.13.3
msrest~=0.7.1
pdf2image~=1.17.0
matplotlib~=3.10.0
opencv-python~=4.11.0.86
pillow~=11.1.0
Flask-Cors~=5.0.0
tiktoken==0.8.0
bcrypt~=4.0.1
pymongo~=4.6.0

# LangChain and LangGraph dependencies
langchain==0.3.12
langchain-openai==0.2.12
langchain-community==0.3.12
langchain-core==0.3.29
langgraph==0.2.60
langchain-chroma==0.1.4
langchain-experimental==0.3.4

# Azure AI dependencies for LangChain
azure-ai-formrecognizer==3.3.0
azure-cognitiveservices-vision-computervision==0.9.0

# Additional dependencies for enhanced functionality
pydantic==2.10.5
tenacity==9.0.0
aiohttp==3.11.11
asyncio==3.4.3
redis==5.2.1